Улучшения кода
