--- victory.py
@@ -1,52 +1,0 @@
-# Ученые: Менделеев Д.И. (08.02.1834), Ломонсов М.В. (19.11.1711), Ковалевская С.Ф. (15.01.1850), Пирогов Н.И. (25.11.1810),
-# Жуковский Н.Е. (17.01.1847), Басов Н.Г. (14.12.1922), Павлов И.П (26.09.1849), Перельман Г.Я. (13.06.1966),
-# Попов А.С. (16.03.1859), Черенков П.А. (28.07.1904)
-
-import random
-
-def victorina():
-    # Входные данные в формате строки
-    born_date = 'Менделеев Д.И.,08.02.1834,Ломонсов М.В.,19.11.1711,Ковалевская С.Ф.,15.01.1850,Пирогов Н.И.,25.11.1810,' \
-                'Жуковский Н.Е.,17.01.1847,Басов Н.Г.,14.12.1922,Павлов И.П.,26.09.1849,Перельман Г.Я.,13.06.1966,Попов А.С.,16.03.1859,Черенков П.А.,28.07.1904'
-    # Создание списка путем разделения строки. Элемент состоит из имени и даты рождения.
-    scientists = born_date.split(',')
-    # Объявление списков для разделения имен и дат рождения
-    scientists_name = []
-    scientists_born_date = []
-    # Цикл добавления имен и дат рождения в разные списки
-    for i in range(len(scientists)):
-        if i % 2 != 0:
-            scientists_born_date.append(scientists[i])
-        else:
-            scientists_name.append(scientists[i])
-    # Создание словаря ключ-имя, значение-дата рождения
-    scientists_dict = dict(zip(scientists_name, scientists_born_date))
-    # Объявление списка с текстовыми датами рождения
-    born_date_txt = ['Восьмое февраля 1834 года', 'Девятнадцатое ноября 1711 года', 'Пятнадцатое января 1850 года',
-                     'двадцать пятое ноября 1810 года', 'семнадцатое января 1847 года', 'четырнадцатое декабря 1922 года',
-                     'Двадцать шестое сентября 1849 года', 'Тринадцатое июня 1966 года', 'Шестнадцатое марта 1859 года',
-                     'Двадцать восмое июля 1904 года']
-    # Создание словаря с текстовыми датами. Ключами являются имена
-    born_date_txt_dict = dict(zip(scientists_name, born_date_txt))
-    # Объявление переменной для входа в цикл
-    cont = 'Да'
-    # Цикл викторины.
-    while cont == 'Да':
-        count_good = 0
-        question_name = random.sample(scientists_name, 5)
-        for j in range(len(question_name)):
-            answer = input(f'Введите дату рождения {question_name[j]} в формате дд.мм.гггг: ')
-            if scientists_dict[question_name[j]] == answer:
-                print('Дата рождения верна')
-                count_good = count_good + 1
-            else:
-                print(f'Дата рождения неверна. Верный ответ: {born_date_txt_dict[question_name[j]]}')
-
-        # Подсчет и вывод количества верных и неверных ответов
-        count_bad = 5 - count_good
-        print(f'Количество верных ответов - {count_good}')
-        print(f'Количество неправильных ответов - {count_bad}')
-        cont = input('Начать викторину занаво? (Да/Нет): ')
-
-if __name__ == '__main__':
-    victorina()
